+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 - [FIXED] Added status code `TDReplicatorErrorNetworkOffline` -
   replicators will go into an error state with this error code if the
   network goes offline, instead of appearing to complete normally.
+- [NEW] Interceptors can return `nil` to signal an error
+  condition. See
+  [the interceptor documentation](doc/httpinterceptors.md) for more
+  details.
 
 ## 1.2.2 (2017-09-06)
 - [IMPROVED] Added pre-emptive session renewal when within 5 minutes of expiry.